+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Suppression du modèle `Utilisateur`, remplacé par `UtilisateurSQL`
 - Suppression des fonctions SAEConnecteur::getAcuseReception et SAEConnecteur::getReply 
 - Suppression de TdtConnecteur::postHelios(), remplacé par TdtConnecteur::sendHelios()
+- Suppression de la méthode magique Controler::__get
 
 # [3.1.13] - 2022-04-18